requirements.txt
django
django-ipware
django-cors-headers
